refactor: reorganize and restructure packages
