--- app/src/main/java/com/android/solvit/model/Language.kt
@@ -1,10 +1,0 @@
-package com.android.solvit.model
-
-enum class Language {
-    ENGLISH,
-    FRENCH,
-    GERMAN,
-    ITALIAN,
-    SPANISH,
-    ARABIC
-}
