--- app/src/main/java/com/android/solvit/model/Services.kt
@@ -1,14 +1,0 @@
-package com.android.solvit.model
-
-enum class Services {
-    PLUMBER,
-    ELECTRICIAN,
-    TUTOR,
-    EVENT_PLANNER,
-    WRITER,
-    CLEANER,
-    CARPENTER,
-    PHOTOGRAPHER,
-    PERSONAL_TRAINER,
-    HAIR_STYLIST
-}
