+++ app/src/main/java/com/android/solvit/model/provider/Language.kt
@@ -1,0 +1,10 @@
+package com.android.solvit.model.provider
+
+enum class Language {
+  ENGLISH,
+  FRENCH,
+  GERMAN,
+  ITALIAN,
+  SPANISH,
+  ARABIC
+}
