+++ app/src/main/java/com/android/solvit/model/provider/Services.kt
@@ -1,0 +1,14 @@
+package com.android.solvit.model.provider
+
+enum class Services {
+  PLUMBER,
+  ELECTRICIAN,
+  TUTOR,
+  EVENT_PLANNER,
+  WRITER,
+  CLEANER,
+  CARPENTER,
+  PHOTOGRAPHER,
+  PERSONAL_TRAINER,
+  HAIR_STYLIST
+}
